Start coding admins

+++ busesProject/src/main/java/busesProject/repositories/UsuarioRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UsuarioRepository extends CrudRepository<Usuario, Integer> {
+public interface UsuarioRepository extends CrudRepository<Usuario, Long> {
 
 
     Optional<Usuario> findByUsername(String username);
